+++ src/main/java/main/Utility.java
@@ -25,13 +25,8 @@ import org.gephi.graph.api.Node;
 import org.gephi.statistics.plugin.EigenvectorCentrality;
 import org.gephi.statistics.plugin.GraphDistance;
 import org.gephi.statistics.plugin.Modularity;
-import org.gephi.statistics.plugin.builder.ConnectedComponentsBuilder;
-import org.gephi.statistics.plugin.builder.ModularityBuilder;
-import org.gephi.statistics.spi.Statistics;
 import org.openide.util.Lookup;
 
-import javax.swing.*;
-import java.awt.print.PrinterException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -162,55 +157,57 @@ public class Utility {
 		}
 
 		nodeLists.add(kNodeList);
- 
+
 		return nodeLists;
 
 	}
 
-	public static float CalculatePopulationScore(String zone, float area) {
+	public static double CalculatePopulationScore(String zone, float area){
 		int density;
 		switch (zone) {
-		case "1":
-			density = 1200;
-			break;
-		case "2A1":
-			density = 5000;
-			break;
-		case "2A":
-			density = 800;
-			break;
-		case "2B":
-		case "3C":
-			density = 600;
-			break;
-		case "2C":
-			density = 350;
-			break;
-		case "3A":
-		case "4A":
-		case "4B":
-			density = 450;
-			break;
-		case "3B":
-			density = 500;
-			break;
-		case "5":
-			density = 350;
-			break;
-		case "6A":
-		case "6A1":
-		case "6B":
-		case "6B1":
-		case "6C":
-		case "6C1":
-			density = 150;
-			break;
-		case "7":
-			density = 200;
-			break;
-		default:
-			return 0;
+			case "1":
+				density = 1200;
+				break;
+			case "2A1":
+				density = 5000;
+				break;
+			case "2A":
+				density = 800;
+				break;
+			case "2B":
+			case "3C":
+				density = 600;
+				break;
+			case "2C":
+				density = 350;
+				break;
+			case "3A":
+			case "4A":
+			case "4B":
+				density = 450;
+				break;
+			case "3B":
+				density = 500;
+				break;
+			case "5":
+				density = 350;
+				break;
+			case "6A":
+			case "6A1":
+			case "6B":
+			case "6B1":
+			case "6C":
+			case "6C1":
+				density = 150;
+				break;
+			case "7":
+				density = 200;
+				break;
+			default:
+				return 0;
 		}
-		return (area / density);
+		return (area/density);
 	}
+
+
 }
